requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-decouple
pydantic
python-multipart
requests
